// src/shared/services/api/models/reviewDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface ReviewDto {
  id?: string;
  user_id?: string;
  /** @nullable */
  user_full_name?: string | null;
  course_id?: string;
  rating?: number;
  /** @nullable */
  review_text?: string | null;
  date_created?: string;
}